server/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six
spacy
openai
pymongo
python-dotenv
